Changed is_integer and is_float to use 64 bit numbers; fixed warning.

+++ src/utility.rs
@@ -1,9 +1,9 @@
 // Tforth utilities
 
 pub fn is_integer(s: &str) -> bool {
-    s.parse::<i32>().is_ok()
+    s.parse::<i64>().is_ok()
 }
 
 pub fn is_float(s: &str) -> bool {
-    s.parse::<f32>().is_ok()
+    s.parse::<f64>().is_ok()
 }
